requirements.txt
paho-mqtt==1.6.1
cryptography==43.0.0
protobuf==5.28.1
